backend/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
python-multipart==0.0.18
pydantic==2.12.3
pillow==10.4.0
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0